requirements.txt
accelerate
dosemetrics
dynamic-network-architectures
hydra-core
matplotlib
monai
numpy
omegaconf
opencv-python
pandas
PyYAML
scipy
seaborn
SimpleITK
torch
torchvision
tqdm
transformers
wandb
